Использовал bindings
Использовал bindings для отображения очков, а так же кнопок начала и
окончания игры.

+++ Tetris/main.m
@@ -12,6 +12,6 @@ int main(int argc, char *argv[])
 {
     time_t t;
     time(&t);
-    srand(t);
+    srand((int)t);
     return NSApplicationMain(argc, (const char **)argv);
 }
